⛵💻 Make sure indices are moved to device (#800)
* add device property to representation module
* move indices to correct device in forward
